@@ -1,8 +1,10 @@
 module github.com/ochinchina/supervisord/config
 
-go 1.16
+go 1.23.0
 
 require (
 	github.com/ochinchina/go-ini v1.0.1
-	github.com/sirupsen/logrus v1.8.1
+	github.com/sirupsen/logrus v1.9.3
 )
+
+require golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
