@@ -1,16 +1,32 @@
 module github.com/ochinchina/supervisord/process
 
-go 1.16
+go 1.23.0
 
 require (
 	github.com/ochinchina/filechangemonitor v0.3.1
-	github.com/ochinchina/supervisord/config v0.0.0-20220721095143-c2527852d28f
-	github.com/ochinchina/supervisord/events v0.0.0-20220721095143-c2527852d28f
-	github.com/ochinchina/supervisord/faults v0.0.0-20220721095143-c2527852d28f // indirect
-	github.com/ochinchina/supervisord/logger v0.0.0-20220721095143-c2527852d28f
-	github.com/ochinchina/supervisord/signals v0.0.0-20220721095143-c2527852d28f
-	github.com/ochinchina/supervisord/util v0.0.0-20220721095143-c2527852d28f // indirect
-	github.com/prometheus/client_golang v1.11.1
+	github.com/ochinchina/supervisord/config v0.0.0-20230902082938-c2cae38b7454
+	github.com/ochinchina/supervisord/events v0.0.0-20230902082938-c2cae38b7454
+	github.com/ochinchina/supervisord/logger v0.0.0-20230902082938-c2cae38b7454
+	github.com/ochinchina/supervisord/signals v0.0.0-20230902082938-c2cae38b7454
+	github.com/prometheus/client_golang v1.20.5
 	github.com/robfig/cron/v3 v3.0.1
-	github.com/sirupsen/logrus v1.8.1
+	github.com/sirupsen/logrus v1.9.3
+)
+
+require (
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.3.0 // indirect
+	github.com/gorilla/rpc v1.2.1 // indirect
+	github.com/hashicorp/go-envparse v0.1.0 // indirect
+	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
+	github.com/ochinchina/go-ini v1.0.1 // indirect
+	github.com/ochinchina/gorilla-xmlrpc v0.0.0-20171012055324-ecf2fe693a2c // indirect
+	github.com/ochinchina/supervisord/faults v0.0.0-20230902082938-c2cae38b7454 // indirect
+	github.com/ochinchina/supervisord/util v0.0.0-20230902082938-c2cae38b7454 // indirect
+	github.com/prometheus/client_model v0.6.1 // indirect
+	github.com/prometheus/common v0.55.0 // indirect
+	github.com/prometheus/procfs v0.15.1 // indirect
+	github.com/rogpeppe/go-charset v0.0.0-20190617161244-0dc95cdf6f31 // indirect
+	golang.org/x/sys v0.27.0 // indirect
+	google.golang.org/protobuf v1.34.2 // indirect
 )
