@@ -1,5 +1,7 @@
 module github.com/ochinchina/supervisord/signals
 
-go 1.16
+go 1.23.0
 
-require github.com/sirupsen/logrus v1.8.1
+require github.com/sirupsen/logrus v1.9.3
+
+require golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
