@@ -117,17 +117,16 @@ func (c *Config) createEntry(name string, configDir string) *Entry {
 	return entry
 }
 
-//
 // Load the configuration and return loaded programs
 func (c *Config) Load() ([]string, error) {
 	myini := ini.NewIni()
 	c.ProgramGroup = NewProcessGroup()
-	log.WithFields(log.Fields{"file": c.configFile}).Info("load configuration from file")
+	log.WithFields(log.Fields{"file": c.configFile}).Info("从文件加载配置")
 	myini.LoadFile(c.configFile)
 
 	includeFiles := c.getIncludeFiles(myini)
 	for _, f := range includeFiles {
-		log.WithFields(log.Fields{"file": f}).Info("load configuration from file")
+		log.WithFields(log.Fields{"file": f}).Info("从文件加载配置")
 		myini.LoadFile(f)
 	}
 	return c.parse(myini), nil
@@ -405,7 +404,8 @@ func parseEnvFiles(s string) *map[string]string {
 }
 
 // GetEnv returns slice of strings with keys separated from values by single "=". An environment string example:
-//  environment = A="env 1",B="this is a test"
+//
+//	environment = A="env 1",B="this is a test"
 func (c *Entry) GetEnv(key string) []string {
 	value, ok := c.keyValues[key]
 	result := make([]string, 0)
@@ -426,7 +426,9 @@ func (c *Entry) GetEnv(key string) []string {
 }
 
 // GetEnvFromFiles returns slice of strings with keys separated from values by single "=". An envFile example:
-//  envFiles = global.env,prod.env
+//
+//	envFiles = global.env,prod.env
+//
 // cat global.env
 // varA=valueA
 func (c *Entry) GetEnvFromFiles(key string) []string {
@@ -512,7 +514,6 @@ func (c *Entry) GetStringArray(key string, sep string) []string {
 //	logSize=1GB
 //	logSize=1KB
 //	logSize=1024
-//
 func (c *Entry) GetBytes(key string, defValue int) int {
 	v, ok := c.keyValues[key]
 
